WIP: altitude extraction (verdict/native modules), ratchet green, full regression batch green
