+++ ouroboros/review_execution.py
@@ -33,8 +33,6 @@ from ouroboros.triad_review import (
     ACCEPTANCE_SURFACE_RULES,
     REVIEW_JSON_ARRAY_CONTRACT,
     TIER_CLASSIFICATION_RULES,
-    empty_array_is_verified_clean,
-    extract_json_array,
 )
 from ouroboros.deadline_utils import (
     bounded_seconds, owner_deadline_exhausted,
@@ -533,196 +531,16 @@ def review_session_output_schema(surface: str) -> Dict[str, Any]:
     shaped["properties"]["findings"]["minItems"] = 1
     return shaped
 
-_UNEXTRACTABLE = "UNEXTRACTABLE"
-
-# The light model canonicalizes NARRATIVE to the review's own output contract —
-# bare ``[]`` or a findings array — so a clean verdict from a session survives
-# exactly as a findings verdict does (D19 closed the asymmetry where a session
-# could BLOCK but never CLEAR). It is the ONE sanctioned second-model use (§8
-# item 6 exception): it extracts; it never judges, repairs, summarizes or
-# attests, and a transcript that is not a completed review comes back
-# UNEXTRACTABLE rather than as an invented verdict.
-_SESSION_EXTRACT_PROMPT = (
-    "The text below is the final answer of a delegated code-review session.\n"
-    "Canonicalize its verdict. Reply with EXACTLY ONE of:\n"
-    "1. [] — when the reviewer COMPLETED the review and explicitly reports no findings\n"
-    "   (a clean verdict). Never reply [] for a refusal, an error, or an unfinished review.\n"
-    "2. ONLY the JSON array of the findings the reviewer reported, copied faithfully\n"
-    "   into the review's own output contract (below). Never invent, merge or drop findings.\n"
-    f"3. The single word {_UNEXTRACTABLE} — when the text is not a completed review\n"
-    "   (a refusal, an error dump, an unfinished session, or anything you cannot map\n"
-    "   faithfully onto the contract).\n"
-    "No prose, no markdown fences.\n\n"
-    "The review's output contract was:\n{contract}\n\n"
-    "Session answer to canonicalize:\n{raw_text}\n"
+# Verdict canonicalization lives in its own module (altitude, P7): both the
+# session route and the native tool-round route consume it.
+from ouroboros.review_verdict_extraction import (  # noqa: F401 — re-exported seam surface
+    _EXTRACT_MAX_CHARS,
+    _UNEXTRACTABLE,
+    _extract_verdict_via_light_model,
+    _findings_array,
+    _strictly_parseable,
+    canonicalize_session_verdict,
 )
-
-# The extraction rail reads the session answer WHOLE — no head/tail window.
-# A windowed read is not a smaller extraction, it is a different (fabricated)
-# verdict: findings reported mid-transcript vanish, and the light model
-# faithfully canonicalizes the cut it was shown into a clean/partial verdict.
-# The single physical send still has one hard bound, because a light model's
-# context is finite: the engine caps a text artifact at 4 MiB while common
-# light-model windows hold ~100k tokens, so past this bound extraction REFUSES
-# with the typed ``extraction_incomplete`` disposition — the raw transcript
-# survives for forensics and can never read as clean — instead of silently
-# shrinking the artifact.
-_EXTRACT_MAX_CHARS = 400_000
-
-
-def _findings_array(payload: Any) -> Optional[List[Dict[str, Any]]]:
-    """The findings list inside a structured payload, or None when it has none."""
-    if isinstance(payload, dict):
-        payload = payload.get("findings")
-    if isinstance(payload, list) and all(isinstance(item, dict) for item in payload):
-        return payload
-    return None
-
-
-def _strictly_parseable(text: str) -> bool:
-    """Would the surfaces' own strict parsers accept this text as a verdict?
-
-    The strict path comes FIRST (D19): a session that already obeyed the output
-    contract is passed through byte-identical, and the constitutional
-    ``empty_array_is_verified_clean`` predicate stays untouched — its strictness
-    is the reason extraction exists, not a defect extraction papers over.
-
-    The WHOLE answer must BE the payload. This used to SCAN with
-    ``extract_json_array``, so any JSON array of objects appearing anywhere in a
-    transcript made it "strict" — a refusal that quoted the contract's own
-    example ("I reviewed NOTHING. The contract asked for entries like
-    [{"item": ..., "verdict": "PASS", ...}]") was passed through byte-identical
-    as a TRUSTED verdict, and the extraction rail that exists precisely to
-    canonicalize a non-verdict never ran. Requiring the whole text removes that
-    leniency; it matches the discipline ``empty_array_is_verified_clean``
-    already applies, and a narrative falls through to extraction, which is what
-    extraction is for.
-    """
-    body = str(text or "")
-    if empty_array_is_verified_clean(body):
-        return True
-    try:
-        parsed = json.loads(body.strip())
-    except (TypeError, ValueError):
-        return False
-    return bool(parsed) and isinstance(parsed, list) and all(isinstance(item, dict) for item in parsed)
-
-
-def canonicalize_session_verdict(
-    raw_text: str, *, conformance_passed: bool, contract: str = "", llm: Any = None,
-    deadline_at: Any = None, transport_timeout_sec: Any = None,
-) -> tuple[str, str, Dict[str, Any]]:
-    """Return ``(canonical_text, method, extraction_usage)`` for a session answer.
-
-    Order is the owner's (D19): trusted structured output first (gated on the
-    run's ``outputConformance == "passed"``, never on run success), then the
-    strict parser, then LIGHT-MODEL extraction over the WHOLE answer.
-    Extraction is not a review call: it runs under its OWN one-send physical
-    rail so it can never consume the reviewing actor's permitted sends, and it
-    spends no reviewer slot. An answer too large for the one-send rail is the
-    typed ``extraction_incomplete`` — never a windowed read, whose canonical
-    form would be a verdict fabricated from the visible cut. ``method`` is one
-    of ``schema | strict | light_model_extraction | extraction_incomplete |
-    unparsed``.
-    """
-    text = str(raw_text or "")
-    if conformance_passed:
-        try:
-            payload = json.loads(text.strip())
-        except (TypeError, ValueError):
-            payload = None
-        findings = _findings_array(payload)
-        if findings is not None:
-            return ("[]" if not findings else json.dumps(findings, ensure_ascii=False)), "schema", {}
-        # The engine claimed conformance over a payload that does not carry the
-        # contract's shape: fall through to the honest branches, and the caller
-        # discloses the delta.
-    if _strictly_parseable(text):
-        return text, "strict", {}
-    if len(text) > _EXTRACT_MAX_CHARS:
-        return text, "extraction_incomplete", {}
-    canonical, usage = _extract_verdict_via_light_model(
-        text, contract=contract, llm=llm, deadline_at=deadline_at,
-        transport_timeout_sec=transport_timeout_sec)
-    if canonical is not None:
-        return canonical, "light_model_extraction", usage
-    # `unparsed` is the honest end of THIS layer's knowledge. The coordinator's
-    # own fenced scanner may still parse the text downstream; labeling that
-    # here would need either a duplicate parser (drift) or a backward import
-    # of the coordinator (the one-way seam ARCHITECTURE pins) — both cost more
-    # than the telemetry cosmetics are worth. Disclosed residual: a fenced
-    # verdict that lands downstream is telemetered `unparsed` at this layer.
-    return text, "unparsed", usage
-
-def _extract_verdict_via_light_model(
-    raw_text: str, *, contract: str = "", llm: Any = None, deadline_at: Any = None,
-    transport_timeout_sec: Any = None) -> tuple[Optional[str], Dict[str, Any]]:
-    """One bounded light-model call canonicalizing narrative to the contract."""
-    from ouroboros.config import get_light_model
-    from ouroboros.usage_accounting import physical_attempt_limit
-
-    if not str(raw_text or "").strip():
-        return None, {}
-    model = get_light_model()
-    from ouroboros.config import get_finalization_grace_sec
-    if owner_deadline_exhausted(deadline_at=deadline_at, reserve_sec=get_finalization_grace_sec()):
-        return None, {"model": model, "reason_code": "deadline_exhausted", "dispatch": "not_dispatched"}
-    prompt = _SESSION_EXTRACT_PROMPT.format(
-        contract=contract or REVIEW_JSON_ARRAY_CONTRACT,
-        raw_text=raw_text,  # WHOLE — the caller already bounded the one send
-    )
-    # Formatting the extraction prompt can itself be expensive. Re-check at
-    # the physical light-model boundary rather than letting an expired owner
-    # window reach the transport helper's positive floor.
-    if owner_deadline_exhausted(deadline_at=deadline_at, reserve_sec=get_finalization_grace_sec()):
-        return None, {"model": model, "reason_code": "deadline_exhausted", "dispatch": "not_dispatched"}
-    try:
-        if llm is None:
-            from ouroboros.llm import LLMClient
-
-            llm = LLMClient()
-        from dataclasses import replace as _replace
-
-        from ouroboros.usage_accounting import UsageScope, current_usage_scope, usage_scope
-
-        # A FRESH one-send rail: the extraction must not claim a send from the
-        # reviewing actor's two-physical-send rail (D19 — not a review call).
-        # The ledger row keeps the actor's task/category attribution but is
-        # sub-labeled `review_substrate.extraction`, so the small light-model
-        # rows beside the $0.00 subscription settlements read as what they are —
-        # verdict extraction, not review-slot spend.
-        _scope = _replace(current_usage_scope() or UsageScope(), source="review_substrate.extraction")
-        chat_kwargs = dict(
-            messages=[{"role": "user", "content": prompt}], model=model,
-            max_tokens=8192, reasoning_effort="low", no_proxy=True,
-        )
-        transport = review_transport_timeout(model, transport_timeout_sec, deadline_at)
-        if transport is not None:
-            chat_kwargs["timeout"] = transport
-        with physical_attempt_limit(1), usage_scope(_scope):
-            message, usage = llm.chat(**chat_kwargs)
-    except Exception as exc:
-        log.warning("Review session verdict extraction failed: %s", exc)
-        return None, {}
-    content = message.get("content") if isinstance(message, dict) else ""
-    if isinstance(content, list):
-        content = " ".join(str(b.get("text", "")) for b in content if isinstance(b, dict))
-    body = str(content or "").strip()
-    usage = dict(usage or {})
-    usage["model"] = model
-    if not body or _UNEXTRACTABLE in body.upper()[:80]:
-        return None, usage
-    if empty_array_is_verified_clean(body):
-        return "[]", usage
-    findings = _findings_array(extract_json_array(body))
-    if findings is None:
-        try:
-            findings = _findings_array(json.loads(body))
-        except (TypeError, ValueError):
-            findings = None
-    if findings is None:
-        return None, usage
-    return ("[]" if not findings else json.dumps(findings, ensure_ascii=False)), usage
 
 
 # ---------------------------------------------------------------------------
@@ -1560,378 +1378,6 @@ class AgentSessionReviewExecutor(ReviewSlotExecutor):
         except Exception:
             log.warning("capability_delta disclosure write failed", exc_info=True)
 
-# ---------------------------------------------------------------------------
-# The native tool-round route variant: a configured-subagent api row RETRIEVES
-# the subject through bounded read-only inspection rounds instead of receiving
-# the assembled packet. Not a third public route kind — ``api_chat`` stays the
-# wire vocabulary; the slot's actor binding selects this executor internally.
-# ---------------------------------------------------------------------------
-
-_INSPECTION_TOOL_NAMES = (
-    "read_file", "list_files", "search_code", "query_code",
-    "vcs_status", "vcs_diff",
-)
-
-_NATIVE_REVIEW_INSTRUCTIONS = (
-    "You are an independent Ouroboros reviewer running a bounded read-only "
-    "inspection episode. Retrieve the evidence yourself with the tools you are "
-    "given inside the repository root: read files, list, search, query code "
-    "structure, and read-only git status/diff. You cannot modify anything and "
-    "have no shell. Work within a hard round budget: read what the checklist "
-    "requires, then answer. Your FINAL message must contain no tool calls and "
-    "must follow the output contract in the task EXACTLY; your host parses it "
-    "structurally, and prose around the verdict is a non-response."
-)
-
-# Physical-send headroom per provider call: llm.py may spend one internal
-# one-shot recovery send per chat() (param-drop/wire-recovery/reroute), and the
-# limit counts PHYSICAL sends. +2 keeps the final-answer round and the
-# coordinator's bookkeeping honest without unbounding anything.
-def native_episode_physical_send_cap() -> int:
-    from ouroboros.config import review_native_max_rounds
-
-    return 2 * review_native_max_rounds() + 2
-
-
-class NativeToolRoundReviewExecutor(ReviewSlotExecutor):
-    """Bounded native inspection episode for a configured-subagent api row.
-
-    ONE episode is ONE logical review attempt: up to the configured round cap
-    of ``LLMClient.chat(tools=…)`` calls against a fresh, instance-local
-    inspection-only ``ToolRegistry``. Every provider call is its own ledger
-    row (the ambient usage scope attributes them); the coordinator's second
-    actor attempt repairs FORMAT locally over the collected answer, exactly
-    like the session executor — there is no mid-episode resume, transcript
-    compaction, or per-round durable ledger, and the caps fail closed.
-    """
-
-    route = ReviewRouteKind.API_CHAT
-
-    def __init__(self, assignment: ReviewAssignment, *, llm: Any = None):
-        super().__init__(assignment, llm=llm)
-        self._episode_prompt: Optional[str] = None
-        self._raw_transcript: Optional[str] = None
-        self._episode_usage: Dict[str, Any] = {}
-        self._tool_receipts: List[Dict[str, Any]] = []
-        self._rounds_used = 0
-        self._settled_failure: Optional[BaseException] = None
-
-    # -- prompt (route-owned; never the api pack) ------------------------------
-
-    def _output_contract(self) -> str:
-        contract = str((self.assignment.request.policy or {}).get("output_contract") or "")
-        return contract or REVIEW_JSON_ARRAY_CONTRACT
-
-    def prompt_payload(self) -> Dict[str, Any]:
-        return {
-            "native_episode_prompt": self.episode_prompt,
-            "inspection_tools": list(_INSPECTION_TOOL_NAMES),
-        }
-
-    def prompt_chars(self) -> int:
-        return len(self.episode_prompt)
-
-    @property
-    def episode_prompt(self) -> str:
-        """The SAME compact work-order the session route sends (D12): task,
-        criteria and output contract, minus any assembled evidence — the actor
-        retrieves context itself, so the packet is never built here."""
-        if self._episode_prompt is None:
-            request, slot = self.assignment.request, self.assignment.slot
-            task = str(request.session_task or "").strip()
-            if not task:
-                raise ReviewRouteUnavailable(
-                    "native tool-round slot has no session task: the surface must "
-                    "supply the route-owned task text (request.session_task) — the "
-                    "assembled api pack is deliberately not sendable to a retrieving "
-                    "actor", code="session_task_missing")
-            parts = [
-                "You are an independent Ouroboros reviewer slot running a bounded "
-                "read-only native inspection episode.",
-                f"Surface: {request.surface}",
-                f"Role hint: {slot.role_hint or 'general reviewer'}",
-                "",
-                task,
-                "",
-                "OUTPUT CONTRACT (your host parses this structurally):",
-                self._output_contract()
-                + "\nYour final message must contain the deliverable alone, with no tool calls.",
-                f"Slot: {slot.slot_id}",
-            ]
-            self._episode_prompt = "\n".join(parts)
-        return self._episode_prompt
-
-    # -- delivery --------------------------------------------------------------
-
-    def execute(self) -> ReviewAttemptResult:
-        if self._raw_transcript is not None:
-            # The permitted resend repairs FORMAT locally over the collected
-            # final answer; it never runs a second paid episode.
-            return self._verdict_result(force_extraction=True)
-        if self._settled_failure is not None:
-            raise self._settled_failure
-        try:
-            self._run_episode()
-        except BaseException as exc:
-            self._settled_failure = exc
-            raise
-        return self._verdict_result()
-
-    def _inspection_registry(self, root: str, scratch: Any) -> tuple[Any, List[Dict[str, Any]]]:
-        """A fresh instance-local registry pinned to ``root``, read-only.
-
-        Reuses the existing capability machinery wholesale instead of a new
-        allowlist mechanism: the ``local_readonly_subagent`` constraint gives
-        the read-only operation/root matrix, ``disabled_tools`` trims its
-        broader name allowlist down to the inspection six, and the resource
-        contract keeps extension/MCP discovery off. ``registry._ctx`` is
-        per-instance, so the worker's own registry/context is never touched.
-        """
-        import pathlib as _pathlib
-
-        from ouroboros.tool_capabilities import LOCAL_READONLY_SUBAGENT_TOOL_NAMES
-        from ouroboros.tools.registry import ToolContext, ToolRegistry
-
-        registry = ToolRegistry(repo_dir=_pathlib.Path(root), drive_root=_pathlib.Path(scratch))
-        ctx = ToolContext(
-            repo_dir=_pathlib.Path(root),
-            drive_root=_pathlib.Path(scratch),
-            task_id=str(self.assignment.request.task_id or "") or None,
-            task_constraint={"mode": "local_readonly_subagent"},
-            task_contract={
-                "allowed_resources": {"network": False, "web": False},
-                "disabled_tools": sorted(
-                    set(LOCAL_READONLY_SUBAGENT_TOOL_NAMES) - set(_INSPECTION_TOOL_NAMES)
-                ),
-            },
-        )
-        registry.set_context(ctx)
-        schemas = []
-        for name in _INSPECTION_TOOL_NAMES:
-            schema = registry.get_schema_by_name(name)
-            if schema:
-                schemas.append(schema)
-        if not schemas:
-            raise ReviewRouteUnavailable(
-                "no inspection tool schemas are projectable for the native "
-                "tool-round episode", code="native_inspection_unavailable")
-        return registry, schemas
-
-    def _run_episode(self) -> None:
-        import shutil
-        import tempfile
-
-        from ouroboros.config import (
-            review_native_max_rounds,
-            review_native_max_transcript_chars,
-        )
-        from ouroboros.llm import add_usage
-        from ouroboros.openai_chat_dispatch import (
-            custom_validation_by_call_id,
-            pop_custom_validation_receipts,
-        )
-
-        request, slot = self.assignment.request, self.assignment.slot
-        root = str(request.session_root or "").strip()
-        if not root:
-            raise ReviewRouteUnavailable(
-                "native tool-round slot has no session root: the surface must name "
-                "the repository root the reviewer episode runs in",
-                code="session_root_missing")
-        chat = getattr(self.llm, "chat", None)
-        if not callable(chat):
-            raise ReviewRouteUnavailable(
-                "native tool-round episode needs a synchronous chat transport",
-                code="api_chat_unavailable")
-        deadline_at = str(getattr(request, "deadline_at", "") or "")
-        rounds_cap = review_native_max_rounds()
-        transcript_cap = review_native_max_transcript_chars()
-        scratch = tempfile.mkdtemp(prefix="ouro-native-review-")
-        registry = None
-        total_usage: Dict[str, Any] = {}
-        transcript_chars = len(self.episode_prompt)
-        final_answer: Optional[str] = None
-        try:
-            registry, schemas = self._inspection_registry(root, scratch)
-            messages: List[Dict[str, Any]] = [
-                {"role": "system", "content": _NATIVE_REVIEW_INSTRUCTIONS},
-                {"role": "user", "content": self.episode_prompt},
-            ]
-            for round_idx in range(1, rounds_cap + 1):
-                if owner_deadline_exhausted(
-                    deadline_at=deadline_at, reserve_sec=get_finalization_grace_sec(),
-                ):
-                    raise _deadline_exhausted_error(
-                        "owner deadline exhausted mid native review episode")
-                chat_kwargs: Dict[str, Any] = {
-                    "messages": messages,
-                    "model": slot.model,
-                    "tools": schemas,
-                    "tool_choice": "auto",
-                    "reasoning_effort": slot.effort,
-                    "max_tokens": int(request.max_tokens or slot.max_tokens),
-                    "no_proxy": bool(request.no_proxy),
-                    "use_local": bool(slot.use_local),
-                    "cache_affinity": f"{request.surface}:{request.task_id or 'review'}",
-                }
-                if request.temperature is not None or slot.temperature is not None:
-                    chat_kwargs["temperature"] = (
-                        request.temperature if request.temperature is not None
-                        else slot.temperature
-                    )
-                transport = review_transport_timeout(
-                    slot.model, getattr(slot, "transport_timeout_sec", None), deadline_at,
-                )
-                if transport is not None:
-                    chat_kwargs["timeout"] = transport
-                with bind_api_review_paid_stamp(self.assignment.dispatch_stamp):
-                    try:
-                        msg, usage = chat(**chat_kwargs)
-                    except BaseException as exc:
-                        capture = getattr(exc, "physical_attempt_capture", None)
-                        if str(getattr(capture, "state", "") or "") in POSITIVE_PHYSICAL_ATTEMPT_STATES:
-                            invoke_review_paid_stamp(self.assignment.dispatch_stamp)
-                        raise
-                self._rounds_used = round_idx
-                tool_calls = (msg.get("tool_calls") or []) if isinstance(msg, dict) else []
-                usage = dict(usage or {})
-                # Pop the wire-validation sidecar BEFORE accumulation, exactly
-                # like the existing bounded loops — receipts are per-round
-                # execution facts, not usage numbers.
-                wire_validation = pop_custom_validation_receipts(usage, tool_calls)
-                validation_by_id = custom_validation_by_call_id(wire_validation)
-                add_usage(total_usage, usage)
-                content = str(msg.get("content") or "") if isinstance(msg, dict) else ""
-                transcript_chars += len(content)
-                if content and not tool_calls:
-                    final_answer = content
-                    break
-                if not tool_calls:
-                    # An empty round is the episode's honest end; the empty
-                    # answer rides the ordinary empty-response rail upstream.
-                    final_answer = content
-                    break
-                assistant = dict(msg)
-                assistant.setdefault("role", "assistant")
-                messages.append(assistant)
-                for tc in tool_calls:
-                    call_id = str(tc.get("id") or "")
-                    function = tc.get("function") if isinstance(tc, dict) else {}
-                    name = str((function or {}).get("name") or "")
-                    raw_args = (function or {}).get("arguments")
-                    args: Optional[Dict[str, Any]] = None
-                    verdict = validation_by_id.get(call_id)
-                    if verdict is not None and not getattr(verdict, "allows_execution", True):
-                        result = f"⚠️ TOOL_ARG_ERROR: {getattr(verdict, 'error', 'invalid arguments')}"
-                    elif name not in _INSPECTION_TOOL_NAMES:
-                        result = (
-                            f"⚠️ tool {name!r} is not available in this read-only "
-                            "inspection episode"
-                        )
-                    else:
-                        try:
-                            args = json.loads(raw_args) if isinstance(raw_args, str) else dict(raw_args or {})
-                            if not isinstance(args, dict):
-                                raise ValueError("arguments must be a JSON object")
-                        except (TypeError, ValueError) as exc:
-                            args, result = None, f"⚠️ TOOL_ARG_ERROR: {exc}"
-                        if isinstance(args, dict):
-                            try:
-                                result = str(registry.execute(name, args))
-                            except Exception as exc:  # tool errors feed the model, not the rail
-                                result = f"⚠️ {type(exc).__name__}: {exc}"
-                    # Host-observed evidence (bounded): which artifacts THIS
-                    # episode actually opened — disclosure, never a claim of
-                    # full-surface coverage.
-                    if len(self._tool_receipts) < 200:
-                        receipt: Dict[str, Any] = {"round": round_idx, "tool": name}
-                        if isinstance(args, dict):
-                            for key in ("path", "root", "query", "pattern"):
-                                if args.get(key):
-                                    receipt[key] = str(args[key])[:300]
-                        receipt["result_chars"] = len(result)
-                        self._tool_receipts.append(receipt)
-                    transcript_chars += len(result)
-                    messages.append({
-                        "role": "tool", "tool_call_id": call_id, "content": result,
-                    })
-                if transcript_chars > transcript_cap:
-                    raise ReviewRouteUnavailable(
-                        f"native review episode transcript ({transcript_chars} chars) "
-                        f"exceeded its bound ({transcript_cap}); the episode fails "
-                        "closed — compaction would review a fabricated cut",
-                        code="native_transcript_cap_exceeded")
-            if final_answer is None:
-                raise ReviewRouteUnavailable(
-                    f"native review episode spent its round budget ({rounds_cap}) "
-                    "without a final answer; the episode fails closed",
-                    code="native_rounds_exhausted")
-        finally:
-            shutil.rmtree(scratch, ignore_errors=True)
-        self._raw_transcript = final_answer
-        self._episode_usage = dict(total_usage)
-        self._episode_usage.update({
-            "provider": str(total_usage.get("provider") or ""),
-            "resolved_model": str(total_usage.get("resolved_model") or slot.model),
-            "native_rounds": self._rounds_used,
-            "native_tool_calls": len(self._tool_receipts),
-            "native_tool_receipts": list(self._tool_receipts),
-            # Provenance class of this delivery: the host SAW these reads.
-            "host_file_read_attestation": "host_observed",
-            "delivery": "native_tool_rounds",
-        })
-
-    def _verdict_result(self, force_extraction: bool = False) -> ReviewAttemptResult:
-        text = self._raw_transcript or ""
-        canonical, method, extraction_usage = canonicalize_session_verdict(
-            text,
-            conformance_passed=False,  # no structured-output channel on this route
-            contract=self._output_contract(),
-            llm=self.llm,
-            deadline_at=getattr(self.assignment.request, "deadline_at", "") or "",
-            transport_timeout_sec=getattr(self.assignment.slot, "transport_timeout_sec", None),
-        )
-        usage = dict(self._episode_usage)
-        deltas: List[Dict[str, Any]] = []
-        if method == "light_model_extraction":
-            usage["extraction"] = extraction_usage
-            deltas.append({
-                "kind": "capability_delta",
-                "requested": "contract-shaped verdict from the episode",
-                "effective": "light-model extraction over the collected answer",
-                "reason": "extraction_instead_of_contract",
-            })
-        elif method == "extraction_incomplete":
-            deltas.append({
-                "kind": "capability_delta",
-                "requested": "contract-shaped verdict from the episode",
-                "effective": (
-                    f"no verdict: answer ({len(text)} chars) exceeds the "
-                    "single-send extraction bound"
-                ),
-                "reason": "extraction_incomplete_transcript_exceeds_bound",
-            })
-        usage["verdict_method"] = method
-        usage["verdict_provenance"] = {
-            "raw_transcript_chars": len(text),
-            "raw_transcript_sha256": hashlib.sha256(text.encode("utf-8", "replace")).hexdigest(),
-            "canonical_chars": len(canonical),
-            "canonical_sha256": hashlib.sha256(canonical.encode("utf-8", "replace")).hexdigest(),
-            "output_conformance": "",
-            "conformance_trusted": False,
-            "verdict_method": method,
-            "raw_transcript_carrier": "message.native_transcript (durable response_ref)",
-        }
-        if deltas:
-            usage["capability_delta"] = deltas
-        message = {
-            "content": canonical,
-            "native_transcript": text,
-            "verdict_method": method,
-        }
-        return ReviewAttemptResult(message=message, usage=usage, raw_text=canonical)
-
-
 # Closed route table. Adding a route means adding an executor here; it never
 # means adding a branch to the coordinator.
 _REVIEW_ROUTE_EXECUTORS: Dict[ReviewRouteKind, type[ReviewSlotExecutor]] = {
@@ -1954,6 +1400,9 @@ def _review_route_executor(assignment: ReviewAssignment, *, llm: Any = None) -> 
     ):
         # A configured-subagent api row is the RETRIEVES class: same wire kind,
         # different delivery — bounded native tool rounds, never the packet.
+        # Imported lazily: the native module subclasses this module's seam.
+        from ouroboros.review_native_episode import NativeToolRoundReviewExecutor
+
         return NativeToolRoundReviewExecutor(assignment, llm=llm)
     executor_cls = _REVIEW_ROUTE_EXECUTORS.get(route)
     if executor_cls is None:
