+++ ouroboros/size_ratchet_manifest.py
@@ -152,6 +152,8 @@ BAND_PATHS = {
     "ouroboros/protected_artifacts.py": None,
     "ouroboros/review.py": "Entered the band from 952 lines: re-anchoring the size ratchet on the official line added the candidate and pairwise base-vs-tip transition validators (validate_size_ratchet_candidate/validate_size_ratchet_transition_against_base) with merge-aware previous-manifest resolution, replacing the retired first-parent history audit (update-flow-redesign sprint, Q7-C/Q18-A/Q19-A owner decisions).",
     "ouroboros/review_custody.py": "Review custody now owns the shared typed retry-rail history and frozen actor reconstruction so physical outcomes cannot be lost between the substrate and reconciliation.",
+    "ouroboros/review_execution.py": "Verdict extraction and the native tool-round executor moved to their own modules (generic reviewer-actor sprint); the transport seam re-enters the band smaller and more cohesive.",
+    "ouroboros/reviewer_slot_config.py": "Absorbed the reviewer_slots() builder from review_substrate (altitude) and configured-subagent row resolution for the generic reviewer-actor bridge.",
     "ouroboros/skill_review_runner.py": None,
     "ouroboros/task_results.py": "Authority reads need an explicit strict mode so malformed child records cannot become a false zero count.",
     "ouroboros/task_status.py": None,
